+++ packages/spruce-event-plugin/src/__tests__/testDirsAndFiles/registered-skill-boot-events/src/listeners/skill/did-boot.v2021_02_08.listener.ts
@@ -6,6 +6,6 @@ import { assert } from '@sprucelabs/test'
 
 export default async (event: SpruceEvent): SpruceEventResponse => {
 	//@ts-ignore
-	assert.isTruthy(event.apiClient)
+	assert.isTruthy(event.mercury)
 	return
 }